refactor: remove draft-service.ts and use API clients directly
- Remove draft-service.ts service layer that was mostly pass-through
- Create draft-utils.ts with utility functions for draft operations
- Update all callsites to use API clients directly via utility functions
- Maintain all existing functionality including email provider drafts
- Preserve data mapping between API responses and internal types
- Improve code organization and reduce unnecessary abstraction

Files changed:
- Deleted: frontend/services/draft-service.ts
- Created: frontend/lib/draft-utils.ts
- Updated: draft-pane.tsx, draft-recovery.tsx, email-view.tsx, use-draft-actions.ts
